Move config and main into erica module
