+++ erica/erica_legacy/elster_xml/transfer_header_fields.py
@@ -1,6 +1,6 @@
 from collections import namedtuple
 
-from erica.erica_legacy.config import get_settings
+from erica.config import get_settings
 
 TransferHeaderFields = namedtuple(
     'TransferHeaderFields',
