+++ tests/infrastructure/rq/test_queue.py
@@ -2,7 +2,7 @@ import pytest
 from redis import Redis
 from rq import Queue
 
-from erica.erica_legacy.config import get_settings
+from erica.config import get_settings
 from erica.infrastructure.rq.queue import get_queue, QueueNotAvailableError
